+++ src/EntityGraphQL/Schema/ArgumentHelper.cs
@@ -65,6 +65,13 @@ namespace EntityGraphQL.Schema
         /// </summary>
         /// <value></value>
         public Expression<Func<TType, bool>> Query { get; set; }
+
+        /// <summary>
+        /// The Compiler will set this based on supplied arguments.null You have use this in your expression to make a choice
+        /// </summary>
+        /// <value></value>
+        public bool HasValue { get; set; }
+
         public EntityQueryType()
         {
             this.QueryType = typeof(TType);