+++ src/EntityGraphQL/Schema/IMethodType.cs
@@ -1,10 +1,17 @@
 using System;
+using System.Collections.Generic;
 
 namespace EntityGraphQL.Schema
 {
     public interface IMethodType
     {
+        IDictionary<string, Type> Arguments { get; }
+        bool IsEnumerable { get; }
+        string Name { get; }
+        Type ReturnTypeClr { get; }
+        string Description { get; }
+
         Type GetArgumentType(string argName);
-        bool HasArgumentType(string argName);
+        bool HasArgumentByName(string argName);
     }
 }